bot.py
from telegram.ext import Updater, CommandHandler
import logging
import openai
from PIL import Image

# Configure logging
logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
logger = logging.getLogger(__name__)

# OpenAI API key
openai.api_key = 'your_openai_api_key_here'

# Telegram Bot Token
TOKEN = 'your_telegram_bot_token_here'

def start(update, context):
    """Send a message when the command /start is issued."""
    update.message.reply_text('Hi! Send /faregpt to generate a branded image.')

def faregpt(update, context):
    # This function will handle the image generation, overlay, and sending process
    update.message.reply_text('Generating your image, please wait...')

    # Step 2.1: Generate an image with DALL-E (placeholder for actual OpenAI call)
    # Note: You need to replace this with actual OpenAI API call to generate an image
    generated_image_path = 'path_to_generated_image.png'  # Placeholder path

    # Step 3: Apply the overlay mask
    # Assuming 'overlay.png' is in the same directory as this script
    final_image_path = apply_overlay(generated_image_path, 'overlay.png')

    # Step 4: Send the final image back to the user
    context.bot.send_photo(chat_id=update.effective_chat.id, photo=open(final_image_path, 'rb'))

def apply_overlay(base_image_path, overlay_image_path):
    # Open the base and overlay images
    base_image = Image.open(base_image_path)
    overlay_image = Image.open(overlay_image_path).resize(base_image.size)

    # Apply the overlay
    base_image.paste(overlay_image, (0, 0), overlay_image)
    final_path = 'final_image.png'
    base_image.save(final_path)
    return final_path

def main():
    """Start the bot."""
    updater = Updater(TOKEN, use_context=True)

    # Get the dispatcher to register handlers
    dp = updater.dispatcher

    # on different commands - answer in Telegram
    dp.add_handler(CommandHandler("start", start))
    dp.add_handler(CommandHandler("faregpt", faregpt))

    # Start the Bot
    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()
