requirements.txt
python-telegram-bot
Pillow
openai
